bondStrategy/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==2.0.12
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
kiwisolver==1.4.4
load-dotenv==0.1.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.5.5
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
typeguard==3.0.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wcwidth==0.2.6
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
